+++ test/support/case.ex
@@ -18,23 +18,21 @@ defmodule Lotus.Case do
   end
 
   setup context do
+    shared = not context[:async]
+
     # Always setup PostgreSQL sandbox (for Lotus storage)
-    pid1 = Ecto.Adapters.SQL.Sandbox.start_owner!(Lotus.Test.Repo, shared: not context[:async])
+    pid1 = Ecto.Adapters.SQL.Sandbox.start_owner!(Lotus.Test.Repo, shared: shared)
     on_exit(fn -> Ecto.Adapters.SQL.Sandbox.stop_owner(pid1) end)
 
     # If test needs SQLite, set up SQLite sandbox too
     if context[:sqlite] do
-      pid2 =
-        Ecto.Adapters.SQL.Sandbox.start_owner!(Lotus.Test.SqliteRepo, shared: not context[:async])
-
+      pid2 = Ecto.Adapters.SQL.Sandbox.start_owner!(Lotus.Test.SqliteRepo, shared: shared)
       on_exit(fn -> Ecto.Adapters.SQL.Sandbox.stop_owner(pid2) end)
     end
 
     # If test needs MySQL, set up MySQL sandbox too
     if context[:mysql] do
-      pid3 =
-        Ecto.Adapters.SQL.Sandbox.start_owner!(Lotus.Test.MysqlRepo, shared: not context[:async])
-
+      pid3 = Ecto.Adapters.SQL.Sandbox.start_owner!(Lotus.Test.MysqlRepo, shared: shared)
       on_exit(fn -> Ecto.Adapters.SQL.Sandbox.stop_owner(pid3) end)
     end