+++ test/support/mysql/migrations/00_create_test_tables.exs
@@ -19,6 +19,7 @@ defmodule Lotus.Test.MysqlRepo.Migrations.CreateTestTables do
       add(:content, :text)
       add(:user_id, references(:test_users, on_delete: :restrict), null: false)
       add(:published, :boolean, default: false)
+      add(:published_at, :utc_datetime)
       add(:view_count, :integer, default: 0)
       add(:tags, :json, default: fragment("('[]')"))
       timestamps()